fix(il/io): validate function header types
Ensure parseFunctionHeader propagates type parse failures and rejects void parameters while keeping return types validated. Update parser unit tests and IL fixtures to use supported types and cover the new diagnostic path.

tests: cmake -S . -B build && cmake --build build && ctest --test-dir build --output-on-failure

+++ tests/unit/test_il_parse_first_error.cpp
@@ -13,7 +13,7 @@
 int main()
 {
     const char *src = R"(il 0.1.2
-func @main() -> i32 {
+func @main() -> i64 {
 entry:
   %0 add 1, 2
   foo %1
