+++ tests/unit/test_il_parse_malformed_func_header.cpp
@@ -12,7 +12,7 @@
 int main()
 {
     const char *src = R"(il 0.1.2
-func @main() -> i32
+func @main() -> i64
 )";
     std::istringstream in(src);
     il::core::Module m;
