+++ tests/unit/test_il_parse_missing_eq.cpp
@@ -12,7 +12,7 @@
 int main()
 {
     const char *src = R"(il 0.1.2
-func @main() -> i32 {
+func @main() -> i64 {
 entry:
   %0 add 1, 2
 }
